Move ensures_aggregatable_across_cases to aggregate.py
