+++ src/CSET/operators/collapse.py
@@ -24,7 +24,7 @@ import iris.util
 
 from CSET._common import iter_maybe
 from CSET.operators._utils import is_time_aggregatable
-from CSET.operators.misc import ensure_aggregatable_across_cases
+from CSET.operators.aggregate import ensure_aggregatable_across_cases
 
 
 def collapse(
