+++ src/CSET/recipes/generic_plevel_histogram_series_case_aggregation_lead_time.yaml
@@ -23,7 +23,7 @@ steps:
         coordinate: pressure
         levels: $PLEVEL
 
-  - operator: misc.ensure_aggregatable_across_cases
+  - operator: aggregate.ensure_aggregatable_across_cases
 
   - operator: plot.plot_histogram_series
     sequence_coordinate: forecast_period
