+++ src/CSET/recipes/generic_surface_histogram_series_case_aggregation_lead_time.yaml
@@ -20,7 +20,7 @@ steps:
         coordinate: pressure
         levels: []
 
-  - operator: misc.ensure_aggregatable_across_cases
+  - operator: aggregate.ensure_aggregatable_across_cases
 
   - operator: write.write_cube_to_nc
     overwrite: True
